requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart